Changed u32->s64 result of VfsEntry::read/write, VfsRamEntry has internal buff and can be read/write

+++ src/filesystem/VfsEntry.h
@@ -42,8 +42,8 @@ public:
 
     // [file interface]
     virtual u32 get_size() const = 0;
-    virtual u32 read(void* data, u32 count) = 0;
-    virtual u32 write(const void* data, u32 count) = 0;
+    virtual s64 read(void* data, u32 count) = 0;
+    virtual s64 write(const void* data, u32 count) = 0;
     virtual bool seek(u32 new_position) = 0;
     virtual bool truncate(u32 new_size) = 0;
